+++ diskord-bot/src/commonMain/kotlin/com/jessecorbett/diskord/bot/BotContext.kt
@@ -5,6 +5,7 @@ import com.jessecorbett.diskord.api.channel.Embed
 import com.jessecorbett.diskord.api.common.Message
 import com.jessecorbett.diskord.api.global.GlobalClient
 import com.jessecorbett.diskord.api.guild.GuildClient
+import com.jessecorbett.diskord.api.interaction.CommandClient
 import com.jessecorbett.diskord.api.interaction.InteractionClient
 import com.jessecorbett.diskord.api.invite.InviteClient
 import com.jessecorbett.diskord.api.webhook.WebhookClient
@@ -45,7 +46,13 @@ public interface BotContext {
     /**
      * Create an instance of an interaction client
      */
-    public fun interaction(applicationId: String): InteractionClient = InteractionClient(applicationId, client)
+    public fun interaction(applicationId: String, interactionToken: String): InteractionClient =
+        InteractionClient(applicationId, interactionToken, client)
+
+    /**
+     * Create an instance of a command client
+     */
+    public fun command(applicationId: String): CommandClient = CommandClient(applicationId, client)
 
     /**
      * Create an instance of a webhook client