+++ diskord-core/src/commonMain/kotlin/com/jessecorbett/diskord/api/interaction/CreateCommand.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-public data class CreateGlobalCommand(
+public data class CreateCommand(
     @SerialName("name") val name: String,
     @SerialName("description") val description: String,
     @SerialName("options") val options: List<CommandOption> = emptyList(),