+++ diskord-core/src/commonMain/kotlin/com/jessecorbett/diskord/api/interaction/InteractionClient.kt
@@ -1,50 +1,80 @@
 package com.jessecorbett.diskord.api.interaction
 
+import com.jessecorbett.diskord.api.common.Message
+import com.jessecorbett.diskord.api.webhook.CreateWebhookMessage
+import com.jessecorbett.diskord.api.webhook.PatchWebhookMessage
 import com.jessecorbett.diskord.internal.client.RestClient
 import com.jessecorbett.diskord.util.DiskordInternals
 import io.ktor.client.call.*
 
 @OptIn(DiskordInternals::class)
-public class InteractionClient(public val applicationId: String, client: RestClient) : RestClient by client {
+public class InteractionClient(
+    public val applicationId: String,
+    public val interactionToken: String,
+    client: RestClient
+) : RestClient by client {
     /**
      * FIXME - Document me!
      */
-    public suspend fun getGlobalCommands(): List<Command> {
-        return GET("/applications/$applicationId/commands").receive()
+    public suspend fun createInteractionResponse(
+        interactionId: String,
+        interactionResponse: InteractionResponse
+    ) {
+        POST("/interactions/$interactionId/$interactionToken/callback") {
+            body = interactionResponse
+        }.receive<Unit>()
     }
 
     /**
      * FIXME - Document me!
      */
-    public suspend fun createGlobalCommand(command: CreateGlobalCommand): Command {
-        return POST("/applications/$applicationId/commands") { body = command }.receive()
+    public suspend fun getOriginalInteractionResponse(): Message {
+        return GET("/webhooks/$applicationId/$interactionToken/messages/@original").receive()
     }
 
     /**
      * FIXME - Document me!
      */
-    public suspend fun getGlobalCommand(commandId: String): Command {
-        return GET("/applications/$applicationId/commands/$commandId").receive()
+    public suspend fun updateOriginalInteractionResponse(message: PatchWebhookMessage): Message {
+        return PATCH("/webhooks/$applicationId/$interactionToken/messages/@original") {
+            body = message
+        }.receive()
     }
 
     /**
      * FIXME - Document me!
      */
-    public suspend fun updateGlobalCommand(commandId: String, command: PatchGlobalCommand): Command {
-        return PATCH("/applications/$applicationId/commands/$commandId") { body = command }.receive()
+    public suspend fun deleteOriginalInteractionResponse() {
+        DELETE("/webhooks/$applicationId/$interactionToken/messages/@original").receive<Unit>()
     }
 
     /**
      * FIXME - Document me!
      */
-    public suspend fun deleteGlobalCommand(commandId: String) {
-        DELETE("/applications/$applicationId/commands/$commandId").receive<Unit>()
+    public suspend fun createFollowupMessage(message: CreateWebhookMessage): Message {
+        return POST("/webhooks/$applicationId/$interactionToken").receive()
     }
 
     /**
      * FIXME - Document me!
      */
-    public suspend fun bulkOverwriteGlobalCommands(commands: List<Command>): List<Command> {
-        return PUT("/applications/$applicationId/commands") { body = commands }.receive()
+    public suspend fun getFollowupMessage(messageId: String): Message {
+        return GET("/webhooks/$applicationId/$interactionToken/messages/$messageId").receive()
+    }
+
+    /**
+     * FIXME - Document me!
+     */
+    public suspend fun updateFollowupMessage(messageId: String, message: PatchWebhookMessage): Message {
+        return PATCH("/webhooks/$applicationId/$interactionToken/messages/$messageId") {
+            body = message
+        }.receive()
+    }
+
+    /**
+     * FIXME - Document me!
+     */
+    public suspend fun deleteFollowupMessage(messageId: String) {
+        DELETE("/webhooks/$applicationId/$interactionToken/messages/$messageId").receive<Unit>()
     }
 }