+++ diskord-core/src/commonMain/kotlin/com/jessecorbett/diskord/api/webhook/CreateWebhookMessage.kt
@@ -2,9 +2,11 @@ package com.jessecorbett.diskord.api.webhook
 
 import com.jessecorbett.diskord.api.channel.AllowedMentions
 import com.jessecorbett.diskord.api.common.Embed
+import com.jessecorbett.diskord.api.common.MessageComponent
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+@Deprecated("Use CreateWebhookMessage instead.", ReplaceWith("CreateWebhookMessage"))
 @Serializable
 public data class WebhookSubmission(
     @SerialName("content") val content: String?,
@@ -15,4 +17,17 @@ public data class WebhookSubmission(
     @SerialName("embeds") val embeds: List<Embed> = emptyList(),
     @SerialName("payload_json") val fileUploadEmbed: String? = null,
     @SerialName("allowed_mentions") val allowedMentions: AllowedMentions = AllowedMentions.ALL
+)
+
+@Serializable
+public data class CreateWebhookMessage(
+    @SerialName("content") val content: String?,
+    @SerialName("username") val overrideUsername: String? = null,
+    @SerialName("avatar_url") val overrideAvatarUrl: String? = null,
+    @SerialName("tts") val useTTS: Boolean = false,
+    @SerialName("file") val fileContent: List<Byte>? = null,
+    @SerialName("embeds") val embeds: List<Embed> = emptyList(),
+    @SerialName("payload_json") val fileUploadEmbed: String? = null,
+    @SerialName("allowed_mentions") val allowedMentions: AllowedMentions = AllowedMentions.ALL,
+    @SerialName("components") val components: List<MessageComponent> = emptyList(),
 )